@@ -1,7 +1,7 @@
 module Spree
   module Admin
     class DigitalsController < ResourceController
-      belongs_to "spree/product", :find_by => :slug
+      belongs_to 'spree/product', find_by: :slug
 
       def create
         invoke_callbacks(:create, :before)
@@ -12,22 +12,23 @@ module Spree
           super
         else
           invoke_callbacks(:create, :fails)
-          flash[:error] = @object.errors.full_messages.join(", ")
+          flash[:error] = @object.errors.full_messages.join(', ')
           redirect_to location_after_save
         end
       end
 
       protected
-        def location_after_save
-          spree.admin_product_digitals_path(@product)
-        end
+
+      def location_after_save
+        spree.admin_product_digitals_path(@product)
+      end
 
       def permitted_resource_params
         params.require(:digital).permit(permitted_digital_attributes)
       end
 
       def permitted_digital_attributes
-        [:variant_id, :attachment, :attachment_file_name, :attachment_content_type]
+        %i[variant_id attachment attachment_file_name attachment_content_type]
       end
 
       def add_additional_paramas
