@@ -2,7 +2,7 @@ module Spree
   module VariantDecorator
     def self.prepended(base)
       base.has_many :digitals
-      base.after_save :destroy_digital, :if => :deleted?
+      base.after_save :destroy_digital, if: :deleted?
     end
 
     # Is this variant to be downloaded by the customer?
@@ -16,7 +16,9 @@ module Spree
     # spree does not delete variants, just marks them as deleted?
     # optionally keep digitals around for customers who require continued access to their purchases
     def destroy_digital
-      digitals.map &:destroy unless Spree::DigitalConfiguration[:keep_digitals]
+      return if Spree::DigitalConfiguration[:keep_digitals]
+
+      digitals.map(&:destroy)
     end
   end
 end
