@@ -2,11 +2,11 @@ module Spree
   module OrderDecorator
     # all products are digital
     def digital?
-      line_items.all? { |item| item.digital? }
+      line_items.all?(&:digital?)
     end
 
     def some_digital?
-      line_items.any? { |item| item.digital? }
+      line_items.any?(&:digital?)
     end
 
     def digital_line_items
@@ -18,9 +18,7 @@ module Spree
     end
 
     def reset_digital_links!
-      digital_links.each do |digital_link|
-        digital_link.reset!
-      end
+      digital_links.each(&:reset!)
     end
   end
 end
