@@ -1,12 +1,11 @@
 require 'spec_helper'
 
 RSpec.describe Spree::Variant do
-
-  context "#destroy" do
+  context '#destroy' do
     let(:variant) { create(:variant) }
     let!(:digital) { create(:digital, variant: variant) }
 
-    it "should destroy associated digitals by default" do
+    it 'should destroy associated digitals by default' do
       # default is false
       Spree::DigitalConfiguration[:keep_digitals] = false
       expect(Spree::Digital.count).to eq(1)
@@ -18,7 +17,7 @@ RSpec.describe Spree::Variant do
       expect(Spree::Digital.count).to eq(0)
     end
 
-    it "should conditionally keep associated digitals" do
+    it 'should conditionally keep associated digitals' do
       Spree::DigitalConfiguration[:keep_digitals] = true
 
       expect(Spree::Digital.count).to eq(1)
